requirements.txt
aniso8601==9.0.1
attrs==23.1.0
autopep8==2.0.2
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
coloredlogs==15.0.1
dnspython==2.3.0
flasgger==0.9.7.1
Flask==2.3.2
Flask-Cors==3.0.10
flask-restplus==0.13.0
humanfriendly==10.0
idna==3.4
ImageHash==4.3.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mistune==2.0.5
mongoengine==0.27.0
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
pycodestyle==2.10.0
pymongo==4.3.3
pyrsistent==0.19.3
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
requests==2.31.0
scipy==1.10.1
six==1.16.0
soupsieve==2.4.1
swagger-ui==0.1.2
terminaltables==3.1.10
tomli==2.0.1
urllib3==2.0.2
Wand==0.6.11
webencodings==0.5.1
Werkzeug==2.3.4
